重构代码
抽取数据库代码为一个类
并且实现一个接口
若换取MyBatis数据库，也可以轻松实现
