+++ .gitignore
@@ -24,3 +24,5 @@ hs_err_pid*
 *.iml
 target/
 .idea/
+news*
+*.db
