+++ src/main/java/com/github/tongkang/JdbcCrawlerDao.java
@@ -1,0 +1,79 @@
+package com.github.tongkang;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class JdbcCrawlerDao implements CrawlerDao{
+
+    private static final String USERNAME = "root";
+    private static final String PASSWORD = "root";
+
+    private final Connection connection;
+
+    public JdbcCrawlerDao() {
+        try {
+            this.connection = DriverManager.getConnection("jdbc:h2:file:D:\\Java_practice\\webCrawler\\news", USERNAME, PASSWORD);
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+
+    public String getNextLink(String sql) throws SQLException {
+        ResultSet resultSet = null;
+        try (PreparedStatement statement = connection.prepareStatement(sql)) {
+            resultSet = statement.executeQuery();
+            while (resultSet.next()) {
+                return resultSet.getString(1);
+            }
+        } finally {
+            if (resultSet != null) {
+                resultSet.close();
+            }
+        }
+        return null;
+    }
+
+    public String getNextLinkThenDelete() throws SQLException {
+        String link = getNextLink("select link from LINKS_TO_BE_PROCESSED LIMIT 1");
+        if (link != null) {
+            updateDatabase(link, "delete from LINKS_TO_BE_PROCESSED where LINK= ?");
+        }
+        return link;
+    }
+
+    public void updateDatabase(String link, String sql) throws SQLException {
+        try (PreparedStatement statement = connection.prepareStatement(sql)) {
+            statement.setString(1, link);
+            statement.executeUpdate();
+        }
+    }
+
+    public void insertNewsIntoDatabase(String url, String title, String content) throws SQLException {
+        try (PreparedStatement statement = connection.prepareStatement("insert into news (url, title, content, created_at,MODIFIED_AT)values(?,?,?,now(),now())")) {
+            statement.setString(1, url);
+            statement.setString(2, title);
+            statement.setString(3, content);
+            statement.executeUpdate();
+        }
+    }
+
+    public boolean isLinkProcessed(String link) throws SQLException {
+        ResultSet resultSet = null;
+        try (PreparedStatement statement = connection.prepareStatement("select LINK from LINKS_ALREADY_PROCESSED where LINK= ?")) {
+            statement.setString(1, link);
+            resultSet = statement.executeQuery();
+            while (resultSet.next()) {
+                return true;
+            }
+        } finally {
+            if (resultSet != null) {
+                resultSet.close();
+            }
+        }
+        return false;
+    }
+}
