--- src/main/java/com/github/tongkang/Main.java
@@ -1,175 +1,0 @@
-package com.github.tongkang;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-
-import java.io.IOException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.stream.Collectors;
-
-public class Main {
-
-    private static final String USERNAME = "root";
-    private static final String PASSWORD = "root";
-
-
-    @SuppressFBWarnings("DMI_CONSTANT_DB_PASSWORD")
-    public static void main(String[] args) throws IOException, SQLException {
-        Connection connection = DriverManager.getConnection("jdbc:h2:file:D:\\Java_practice\\webCrawler\\news", USERNAME, PASSWORD);
-
-        String link;
-
-        //从数据库中加载下一个链接，如果能加载到，就进行循环
-        while ((link = getNextLinkThenDelete(connection)) != null) {
-
-            //询问数据库，当前链接是不是已经被处理过了？
-            if (isLinkProcessed(connection, link)) {
-                continue;
-            }
-
-            if (isInterestingLink(link)) {
-                System.out.println(link);
-                Document doc = httpGetAndParseHtml(link);
-
-                //获取所有的链接
-                parseUrlsFromPageAndStoreIntoDatabase(connection, doc);
-
-                storeIntoDatabaseIfNewsPage(connection, doc, link);
-
-                //加入数据库
-                updateDatabase(connection, link, "insert into LINKS_ALREADY_PROCESSED (link) values ?");
-
-
-            }
-        }
-
-
-    }
-
-    private static String getNextLinkThenDelete(Connection connection) throws SQLException {
-        String link = getNextLink(connection, "select link from LINKS_TO_BE_PROCESSED LIMIT 1");
-        if (link != null) {
-            updateDatabase(connection, link, "delete from LINKS_TO_BE_PROCESSED where LINK= ?");
-        }
-        return link;
-    }
-
-    private static void parseUrlsFromPageAndStoreIntoDatabase(Connection connection, Document doc) throws SQLException {
-        for (Element aTag : doc.select("a")) {
-            String href = aTag.attr("href");
-            if (href.startsWith("//")) {
-                href = "https:" + href;
-            }
-            if (!href.toLowerCase().startsWith("javascript")) {
-                updateDatabase(connection, href, "insert into LINKS_TO_BE_PROCESSED (link) values ?");
-            }
-        }
-    }
-
-    private static boolean isLinkProcessed(Connection connection, String link) throws SQLException {
-        ResultSet resultSet = null;
-        try (PreparedStatement statement = connection.prepareStatement("select LINK from LINKS_ALREADY_PROCESSED where LINK= ?")) {
-            statement.setString(1, link);
-            resultSet = statement.executeQuery();
-            while (resultSet.next()) {
-                return true;
-            }
-        } finally {
-            if (resultSet != null) {
-                resultSet.close();
-            }
-        }
-        return false;
-    }
-
-    private static void updateDatabase(Connection connection, String link, String sql) throws SQLException {
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, link);
-            statement.executeUpdate();
-        }
-    }
-
-    private static String getNextLink(Connection connection, String sql) throws SQLException {
-        ResultSet resultSet = null;
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            resultSet = statement.executeQuery();
-            while (resultSet.next()) {
-                return resultSet.getString(1);
-            }
-        } finally {
-            if (resultSet != null) {
-                resultSet.close();
-            }
-        }
-        return null;
-    }
-
-    private static void storeIntoDatabaseIfNewsPage(Connection connection, Document doc, String link) throws SQLException {
-        //假如这是一个新闻的详情页面，就存入数据库，否则，就什么都不做
-        ArrayList<Element> articleTags = doc.select("article");
-        if (!articleTags.isEmpty()) {
-            for (Element articleTag : articleTags) {
-                String title = articleTags.get(0).child(0).text();
-                String content = articleTag.select("p").stream().map(Element::text).collect(Collectors.joining("\n"));
-                try (PreparedStatement statement = connection.prepareStatement("insert into news (url, title, content, created_at,MODIFIED_AT)values(?,?,?,now(),now())")) {
-                    statement.setString(1, link);
-                    statement.setString(2, title);
-                    statement.setString(3, content);
-                    statement.executeUpdate();
-                }
-            }
-        }
-    }
-
-    @SuppressFBWarnings("RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE")
-    private static Document httpGetAndParseHtml(String link) throws IOException {
-        //这是我们感兴趣的，我们需要处理它
-        CloseableHttpClient httpclient = HttpClients.createDefault();
-
-
-        HttpGet httpGet = new HttpGet(link);
-        httpGet.addHeader("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.90 Safari/537.36");
-
-        try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-            HttpEntity entity1 = response1.getEntity();
-            String html = EntityUtils.toString(entity1);
-
-            return Jsoup.parse(html);
-        }
-
-    }
-
-    private static boolean isInterestingLink(String link) {
-        return (isNewPage(link) || isIndexPage(link)) && isNotLoginPage(link) && isNotGamePage(link);
-    }
-
-    private static boolean isIndexPage(String link) {
-        return "https://sina.cn".equals(link);
-    }
-
-    private static boolean isNewPage(String link) {
-        return link.contains("sina.cn");
-    }
-
-    private static boolean isNotLoginPage(String link) {
-        return !link.contains("passport.sina.cn");
-    }
-
-    private static boolean isNotGamePage(String link) {
-        return !link.contains("games.sina.cn") && !link.contains("lives.sina.cn") && !link.contains("weather1.sina.cn");
-    }
-
-}
